move-Button-component-to-common-folder relate #48
