+++ client/src/Component/Common/BannerFooter/index.js
@@ -3,7 +3,7 @@ import PropTypes from 'prop-types';
 import ReactRouterPropTypes from 'react-router-prop-types';
 import { Link } from 'react-router-dom';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
-import Button from '../../Button';
+import Button from '../Button';
 import {
   BannerFooterWrapper,
   ArrowIcon,
